etl/requirements.txt
# Airflow 3
apache-airflow==3.1.0
apache-airflow-providers-mysql==5.6.3
apache-airflow-providers-celery==3.13.0

# Celery & Redis
celery[redis]>=5.5.0
redis>=5.2.0

# Database
mysql-connector-python==9.1.0

# Web Scraping
beautifulsoup4==4.12.3
requests==2.32.3
lxml==5.3.0
fake-useragent==1.5.1

# Data Processing
pandas==2.2.3
python-dateutil==2.9.0

# Telegram
python-telegram-bot==21.7

# AI/Translation
openai==1.55.3

# Utilities
pydantic>=2.11.0
python-dotenv==1.0.1

